fix: extract plo trend sparkline module
- move sparkline rendering out of plo_trend.js to clear code-sprawl at the root
- harden injectSparklines against selector-breaking ids
- add regression coverage for escaped ids after extraction
