+++ .secrets.baseline
@@ -143,5 +143,5 @@
       }
     ]
   },
-  "generated_at": "2026-04-03T19:48:49Z"
+  "generated_at": "2026-04-03T20:15:13Z"
 }
