+++ templates/plo_dashboard.html
@@ -175,6 +175,7 @@
         crossorigin="anonymous"></script>
 <script src="{{ url_for('static', filename='dashboard_utils.js') }}"></script>
 <script src="{{ url_for('static', filename='plo_trend_panel.js') }}"></script>
+<script src="{{ url_for('static', filename='plo_trend_sparkline.js') }}"></script>
 <script src="{{ url_for('static', filename='plo_detail_panel.js') }}"></script>
 <script src="{{ url_for('static', filename='plo_narrative_progress.js') }}"></script>
 <script src="{{ url_for('static', filename='plo_trend.js') }}"></script>
